+++ src/main/java/org/example/survey/service/CategoryService.java
@@ -1,16 +1,16 @@
 package org.example.survey.service;
 
 import org.example.survey.dto.CategoryDto;
-import org.example.survey.exception.user.RessourceNotFoundException;
+import org.example.survey.exception.RessourceNotFoundException;
 
 import java.util.List;
 import java.util.Optional;
 
 public interface CategoryService {
      Optional<CategoryDto>  add(CategoryDto categoryRequestDto);
-     Optional<CategoryDto> updatePartially(Integer categoryId ,CategoryDto categoryRequestDto) throws RessourceNotFoundException;
-     Optional<CategoryDto> update(Integer categoryId ,CategoryDto categoryRequestDto) throws RessourceNotFoundException;
-     void deleteById(Integer id) throws RessourceNotFoundException;;
+     Optional<CategoryDto> updatePartially(Long categoryId ,CategoryDto categoryRequestDto) throws RessourceNotFoundException;
+     Optional<CategoryDto> update(Long categoryId ,CategoryDto categoryRequestDto) throws RessourceNotFoundException;
+     void deleteById(Long id) throws RessourceNotFoundException;;
      List<CategoryDto> displayAll();
-     Optional<CategoryDto> findByCategoryId(Integer categoryId) throws RessourceNotFoundException;
+     Optional<CategoryDto> findByCategoryId(Long categoryId) throws RessourceNotFoundException;
 }
